refactor helper name
